requirements.txt
robin_stocks==2.1.0
pandas==2.1.0
numpy==1.24.3
flask==2.3.3
werkzeug>=2.3.7
Pillow==9.5.0
google-generativeai==0.3.1
python-dotenv==1.0.0